\name{AIS}
\docType{data}
\alias{AIS}
\title{AIS data}
\description{
The set of AIS data involves recorded body factors of 202 athletes including 100 women 102 men, see Cook (2009). Among factors, two variables body mass index (BMI) and body fat percentage (Bfat) are chosen for cluster analysis.}
\usage{data(AIS)}
\format{A text file with 3 columns.}
\encoding{latin1}
\references{
R. D. Cook and S. Weisberg, (2009). \emph{An Introduction to Regression Graphics}, John Wiley & Sons, New York.
}
\keyword{datasets}
\examples{
data(AIS)
}
\keyword{datasets}

