\name{bankruptcy}
\docType{data}
\alias{bankruptcy}
\title{bankruptcy data}
\description{
The bankruptcy dataset involves ratio of the retained earnings (RE) to the total assets, and the ratio of earnings before interests and the taxes (EBIT) to the total assets of 66 American firms, see Altman (1969).
}
\usage{data(bankruptcy)}
\format{A text file with 3 columns.}
\encoding{latin1}
\references{
E. I. Altman, 1969. Financial ratios, discriminant analysis and the prediction of corporate bankruptcy, \emph{The Journal of Finance}, 23(4), 589-609.
}
\keyword{datasets}
\examples{
data(bankruptcy)
}
