\name{iris}
\docType{data}
\alias{iris}
\title{iris data}
\description{
This set of data introduced by Ronald Fisher (1936). It contains three plant species (setosa, virginica, versicolor) and four features measured for each sample. These quantify the morphologic variation of the iris flower in its three species. All measurements are given in centimeters.
}
\usage{data(iris)}
\format{A text file with 3 columns.}
\encoding{latin1}
\references{
R. A. Fisher, (1936). The use of multiple measurements in taxonomic problems, \emph{Annals of eugenics}, 7(2), 179-188.
}
\keyword{datasets}
\examples{
data(iris)
}
\keyword{datasets}
