\name{wheat}
\docType{data}
\alias{wheat}
\title{wheat data}
\description{
These data are about 210 wheat grains belonging to three different varieties (including: Kama, Rosa, and Canadian) on which 7 quantitative variables related to these kernel structures detected by using a soft X-ray visualization technique have been measured. These variables are: \code{area}, \code{perimeter}, \code{compactness}, \code{length of kernel}, \code{width of kernel}, \code{asymmetry coefficient}, \code{length of kernel groove}, and class label variable \code{variety}.
}
\usage{data(wheat)}
\format{A text file with 8 columns.}
\encoding{latin1}
\references{
P. Giordani, M. B. Ferraro and F. Martella, (2020). \emph{An Introduction to Clustering with R}, Springer, Singapore.
}
\keyword{datasets}
\examples{
data(wheat)
}
\keyword{datasets}

